+++ internal/cli/common/run_flags.go
@@ -138,11 +138,6 @@ func RunFlags(opts *CLIOpts, hidden bool) []cli.Flag {
 			Usage:   "EXPERIMENTAL: watch config files for changes and automatically apply them",
 		},
 	}
-	if hidden {
-		return f
-	}
-
-	// Only add custom flags when not hidden
 	return append(f, opts.CustomRunFlags...)
 }
 
@@ -194,7 +189,7 @@ func PreApplyEnvFilesAndTemplates(c *cli.Context, opts *CLIOpts) error {
 	if err != nil {
 		return fmt.Errorf("failed to resolve template glob pattern: %w", err)
 	}
-	lints, err := template.InitTemplates(templatesPaths...)
+	lints, err := template.InitTemplates(opts.Environment, opts.BloblEnvironment, templatesPaths...)
 	if err != nil {
 		return fmt.Errorf("template file read error: %w", err)
 	}